Move Edge class to core package, next to SequenceNode
